Introduction to Test Driven Development (TDD)/labs/05_factories_and_fakes/requirements.txt
# Dependencies for this lab
Flask
Flask-SQLAlchemy
nose
pinocchio
coverage
factory-boy
pylint
